fix(core): keep Harness state schema in Session
